Change the type of counter from int to size_t

+++ Newton/src/functions/quadratic_function.c
@@ -10,7 +10,7 @@
 void quadratic_function(__attribute__((unused)) void *empty, double *x, size_t size_of_pb, double *func,
                         double *dfunc)
 {
-    for (int i = 0; i < size_of_pb; ++i)
+    for (size_t i = 0; i < size_of_pb; ++i)
     {
         func[i] = x[i] * x[i] + 2. * x[i] - 9.0;
         dfunc[i] = 2 * x[i] + 2.;
